refactor: extract Stake component logic and modals
